Rename and refactor files and create common-canvas.ts and commproperties.ts
