+++ canvas_modules/common-canvas/types/schema-types/palette-v3.ts
